fix(mattermost): preserve immediate edit ordering
