+++ extensions/mattermost/src/mattermost/monitor.inbound-system-event.test-helper.ts
@@ -1,0 +1,74 @@
+// Shared fake websocket for Mattermost inbound monitor tests.
+export class FakeWebSocket {
+  public readonly sent: string[] = [];
+  private readonly openListeners: Array<() => void> = [];
+  private readonly messageListeners: Array<(data: Buffer) => void | Promise<void>> = [];
+  private readonly pongListeners: Array<(data: Buffer) => void> = [];
+  private readonly closeListeners: Array<(code: number, reason: Buffer) => void> = [];
+  private readonly errorListeners: Array<(err: unknown) => void> = [];
+
+  on(event: "open", listener: () => void): void;
+  on(event: "message", listener: (data: Buffer) => void | Promise<void>): void;
+  on(event: "pong", listener: (data: Buffer) => void): void;
+  on(event: "close", listener: (code: number, reason: Buffer) => void): void;
+  on(event: "error", listener: (err: unknown) => void): void;
+  on(event: "open" | "message" | "pong" | "close" | "error", listener: unknown): void {
+    if (event === "open") {
+      this.openListeners.push(listener as () => void);
+      return;
+    }
+    if (event === "message") {
+      this.messageListeners.push(listener as (data: Buffer) => void | Promise<void>);
+      return;
+    }
+    if (event === "pong") {
+      this.pongListeners.push(listener as (data: Buffer) => void);
+      return;
+    }
+    if (event === "close") {
+      this.closeListeners.push(listener as (code: number, reason: Buffer) => void);
+      return;
+    }
+    this.errorListeners.push(listener as (err: unknown) => void);
+  }
+
+  send(data: string): void {
+    this.sent.push(data);
+  }
+
+  ping(): void {}
+
+  close(): void {}
+
+  terminate(): void {
+    this.emitClose(1000);
+  }
+
+  get openListenerCount(): number {
+    return this.openListeners.length;
+  }
+
+  emitOpen(): void {
+    for (const listener of this.openListeners) {
+      listener();
+    }
+  }
+
+  async emitMessage(payload: unknown): Promise<void> {
+    const buffer = Buffer.from(JSON.stringify(payload), "utf8");
+    await Promise.all(this.messageListeners.map((listener) => Promise.resolve(listener(buffer))));
+  }
+
+  emitClose(code: number, reason = ""): void {
+    const buffer = Buffer.from(reason, "utf8");
+    for (const listener of this.closeListeners) {
+      listener(code, buffer);
+    }
+  }
+
+  emitError(err: unknown): void {
+    for (const listener of this.errorListeners) {
+      listener(err);
+    }
+  }
+}
