+++ extensions/mattermost/src/mattermost/monitor.inbound-system-event.test.ts
@@ -1,523 +1,24 @@
 // Mattermost tests cover monitor.inbound system event plugin behavior.
 import { createInboundDebouncer } from "openclaw/plugin-sdk/channel-inbound-debounce";
 import { beforeEach, describe, expect, it, vi } from "vitest";
+import { FakeWebSocket } from "./monitor.inbound-system-event.test-helper.js";
+import {
+  createRuntimeCore,
+  emitMattermostChannelPost,
+  getMattermostInboundTestState,
+  mentionConfig,
+  resetMattermostInboundTestState,
+  testConfig,
+  testRuntime,
+} from "./monitor.inbound-system-event.test-support.js";
+// Load the shared mocks before importing the monitor under test.
 import { monitorMattermostProvider } from "./monitor.js";
-import type { OpenClawConfig, ReplyPayload, RuntimeEnv } from "./runtime-api.js";
+import type { OpenClawConfig } from "./runtime-api.js";
 
-class FakeWebSocket {
-  public readonly sent: string[] = [];
-  private readonly openListeners: Array<() => void> = [];
-  private readonly messageListeners: Array<(data: Buffer) => void | Promise<void>> = [];
-  private readonly pongListeners: Array<(data: Buffer) => void> = [];
-  private readonly closeListeners: Array<(code: number, reason: Buffer) => void> = [];
-  private readonly errorListeners: Array<(err: unknown) => void> = [];
-
-  on(event: "open", listener: () => void): void;
-  on(event: "message", listener: (data: Buffer) => void | Promise<void>): void;
-  on(event: "pong", listener: (data: Buffer) => void): void;
-  on(event: "close", listener: (code: number, reason: Buffer) => void): void;
-  on(event: "error", listener: (err: unknown) => void): void;
-  on(event: "open" | "message" | "pong" | "close" | "error", listener: unknown): void {
-    if (event === "open") {
-      this.openListeners.push(listener as () => void);
-      return;
-    }
-    if (event === "message") {
-      this.messageListeners.push(listener as (data: Buffer) => void | Promise<void>);
-      return;
-    }
-    if (event === "pong") {
-      this.pongListeners.push(listener as (data: Buffer) => void);
-      return;
-    }
-    if (event === "close") {
-      this.closeListeners.push(listener as (code: number, reason: Buffer) => void);
-      return;
-    }
-    this.errorListeners.push(listener as (err: unknown) => void);
-  }
-
-  send(data: string): void {
-    this.sent.push(data);
-  }
-
-  ping(): void {}
-
-  close(): void {}
-
-  terminate(): void {
-    this.emitClose(1000);
-  }
-
-  get openListenerCount(): number {
-    return this.openListeners.length;
-  }
-
-  emitOpen(): void {
-    for (const listener of this.openListeners) {
-      listener();
-    }
-  }
-
-  async emitMessage(payload: unknown): Promise<void> {
-    const buffer = Buffer.from(JSON.stringify(payload), "utf8");
-    await Promise.all(this.messageListeners.map((listener) => Promise.resolve(listener(buffer))));
-  }
-
-  emitClose(code: number, reason = ""): void {
-    const buffer = Buffer.from(reason, "utf8");
-    for (const listener of this.closeListeners) {
-      listener(code, buffer);
-    }
-  }
-
-  emitError(err: unknown): void {
-    for (const listener of this.errorListeners) {
-      listener(err);
-    }
-  }
-}
-
-const mockState = vi.hoisted(() => ({
-  abortController: undefined as AbortController | undefined,
-  createReplyDispatcherWithTyping: vi.fn(),
-  createMattermostClient: vi.fn(),
-  createMattermostDraftStream: vi.fn(),
-  dispatchInboundMessage: vi.fn(),
-  enqueueSystemEvent: vi.fn(),
-  fetchMattermostMe: vi.fn(),
-  registerMattermostMonitorSlashCommands: vi.fn(),
-  registerPluginHttpRoute: vi.fn(),
-  recordMattermostThreadParticipation: vi.fn(),
-  resolveChannelInfo: vi.fn(),
-  resolveMattermostMedia: vi.fn(),
-  resolveUserInfo: vi.fn(),
-  runtimeCore: undefined as unknown,
-  sendMessageMattermost: vi.fn(),
-  updateMattermostPost: vi.fn(),
-}));
-
-vi.mock("openclaw/plugin-sdk/reply-runtime", async (importOriginal) => {
-  const actual = await importOriginal<typeof import("openclaw/plugin-sdk/reply-runtime")>();
-  return {
-    ...actual,
-    createReplyDispatcherWithTyping: (...args: unknown[]) =>
-      mockState.createReplyDispatcherWithTyping(...args),
-    dispatchInboundMessage: async (params: Parameters<typeof actual.dispatchInboundMessage>[0]) => {
-      try {
-        return await mockState.dispatchInboundMessage(params);
-      } finally {
-        await params.onSettled?.();
-      }
-    },
-  };
-});
-
-vi.mock("./client.js", async () => {
-  const actual = await vi.importActual<typeof import("./client.js")>("./client.js");
-  return {
-    ...actual,
-    createMattermostClient: mockState.createMattermostClient,
-    fetchMattermostMe: mockState.fetchMattermostMe,
-    normalizeMattermostBaseUrl: (value: string | undefined) => value?.trim() ?? "",
-    updateMattermostPost: mockState.updateMattermostPost,
-  };
-});
-
-vi.mock("./draft-stream.js", async () => {
-  const actual = await vi.importActual<typeof import("./draft-stream.js")>("./draft-stream.js");
-  return {
-    createMattermostDraftStream: mockState.createMattermostDraftStream,
-    createMattermostDraftPreviewBoundaryController:
-      actual.createMattermostDraftPreviewBoundaryController,
-  };
-});
-
-vi.mock("./monitor-resources.js", async (importOriginal) => ({
-  ...(await importOriginal<typeof import("./monitor-resources.js")>()),
-  createMattermostMonitorResources: () => ({
-    resolveMattermostMedia: mockState.resolveMattermostMedia,
-    sendTypingIndicator: vi.fn(async () => {}),
-    resolveChannelInfo: mockState.resolveChannelInfo,
-    resolveUserInfo: mockState.resolveUserInfo,
-    updateModelPickerPost: vi.fn(async () => {}),
-  }),
-}));
-
-vi.mock("./monitor-slash.js", () => ({
-  registerMattermostMonitorSlashCommands: mockState.registerMattermostMonitorSlashCommands,
-}));
-
-vi.mock("./thread-participation.js", async (importOriginal) => ({
-  ...(await importOriginal<typeof import("./thread-participation.js")>()),
-  recordMattermostThreadParticipation: mockState.recordMattermostThreadParticipation,
-}));
-
-vi.mock("./runtime-api.js", async () => {
-  const actual = await vi.importActual<typeof import("./runtime-api.js")>("./runtime-api.js");
-  return {
-    ...actual,
-    buildAgentMediaPayload: vi.fn(() => ({})),
-    createChannelPairingController: vi.fn(() => ({
-      readStoreForDmPolicy: vi.fn(async () => []),
-      upsertPairingRequest: vi.fn(async () => ({ code: "123456", created: true })),
-    })),
-    createChannelMessageReplyPipeline: vi.fn((params: { cfg: OpenClawConfig }) => ({
-      onModelSelected: vi.fn(),
-      typingCallbacks: {},
-      resolveResponsePrefix: () => params.cfg.messages?.responsePrefix,
-    })),
-    registerPluginHttpRoute: mockState.registerPluginHttpRoute,
-    resolveChannelMediaMaxBytes: vi.fn(() => 8 * 1024 * 1024),
-    warnMissingProviderGroupPolicyFallbackOnce: vi.fn(),
-  };
-});
-
-vi.mock("./send.js", async () => {
-  const actual = await vi.importActual<typeof import("./send.js")>("./send.js");
-  return {
-    ...actual,
-    sendMessageMattermost: mockState.sendMessageMattermost,
-  };
-});
-
-function createRuntimeCore(
-  cfg: OpenClawConfig,
-  routeOverride?: {
-    accountId?: string;
-    agentId?: string;
-    lastRoutePolicy?: "main" | "session";
-    mainSessionKey?: string;
-    sessionKey?: string;
-  },
-  overrides: {
-    inboundDebounceMs?: number;
-    isControlCommandMessage?: (text?: string) => boolean;
-    shouldComputeCommandAuthorized?: (text?: string) => boolean;
-    shouldHandleTextCommands?: () => boolean;
-    textHasControlCommand?: (text?: string) => boolean;
-    createInboundDebouncer?: typeof createInboundDebouncer;
-    verboseDebug?: (message: string) => void;
-    chunkMarkdownTextWithMode?: (
-      text: string,
-      limit: number,
-      mode: "length" | "newline",
-    ) => string[];
-    chunkMode?: "length" | "newline";
-    textChunkLimit?: number;
-  } = {},
-) {
-  type ReplyDispatcherOptions = {
-    deliver: (payload: ReplyPayload, info: { kind: "tool" | "block" | "final" }) => Promise<void>;
-  };
-  mockState.createReplyDispatcherWithTyping.mockImplementation(
-    (options: ReplyDispatcherOptions) => ({
-      dispatcher: {},
-      replyOptions: {},
-      markDispatchIdle: vi.fn(),
-      markRunComplete: vi.fn(),
-      options,
-    }),
-  );
-  type RecordInboundSessionInput = {
-    storePath: string;
-    sessionKey: string;
-    ctx: unknown;
-    createIfMissing?: boolean;
-    groupResolution?: unknown;
-    onRecordError?: (error: unknown) => void;
-    updateLastRoute?: {
-      accountId?: string;
-      channel?: string;
-      mainDmOwnerPin?: {
-        onSkip?: () => void;
-        ownerRecipient?: string;
-        senderRecipient?: string;
-      };
-      sessionKey?: string;
-      to?: string;
-    };
-  };
-  const recordInboundSession = vi.fn(async (_params: RecordInboundSessionInput) => {});
-  const dispatchPreparedForTest = vi.fn(
-    async (turn: {
-      route: { agentId: string; sessionKey: string };
-      ctxPayload: { SessionKey?: string };
-      record?: {
-        groupResolution?: unknown;
-        createIfMissing?: boolean;
-        updateLastRoute?: RecordInboundSessionInput["updateLastRoute"];
-        onRecordError?: (err: unknown) => void;
-      };
-      runDispatch: () => Promise<{
-        queuedFinal: boolean;
-        counts: { tool: number; block: number; final: number };
-      }>;
-    }) => {
-      await recordInboundSession({
-        storePath: "/tmp/openclaw-test-sessions.json",
-        sessionKey: turn.ctxPayload.SessionKey ?? turn.route.sessionKey,
-        ctx: turn.ctxPayload,
-        groupResolution: turn.record?.groupResolution,
-        createIfMissing: turn.record?.createIfMissing,
-        updateLastRoute: turn.record?.updateLastRoute,
-        onRecordError: turn.record?.onRecordError ?? (() => undefined),
-      });
-      const dispatchResult = await turn.runDispatch();
-      return {
-        admission: { kind: "dispatch" as const },
-        dispatched: true,
-        ctxPayload: turn.ctxPayload,
-        routeSessionKey: turn.route.sessionKey,
-        dispatchResult,
-      };
-    },
-  );
-  const run = vi.fn(
-    async (params: {
-      raw: unknown;
-      adapter: {
-        ingest: (raw: unknown) => unknown;
-        resolveTurn: (
-          input: unknown,
-          eventClass: { kind: "message"; canStartAgentTurn: true },
-          preflight: Record<string, never>,
-        ) => Parameters<typeof dispatchPreparedForTest>[0];
-      };
-    }) => {
-      const input = params.adapter.ingest(params.raw);
-      const turn = params.adapter.resolveTurn(
-        input,
-        { kind: "message", canStartAgentTurn: true },
-        {},
-      );
-      return await dispatchPreparedForTest(turn);
-    },
-  );
-  return {
-    config: {
-      current: () => cfg,
-    },
-    logging: {
-      shouldLogVerbose: () => Boolean(overrides.verboseDebug),
-      getChildLogger: () => ({
-        debug: overrides.verboseDebug ?? vi.fn(),
-        info: vi.fn(),
-        warn: vi.fn(),
-        error: vi.fn(),
-      }),
-    },
-    media: {
-      mediaKindFromMime: () => "document",
-    },
-    system: {
-      enqueueSystemEvent: mockState.enqueueSystemEvent,
-    },
-    channel: {
-      activity: {
-        record: vi.fn(),
-      },
-      commands: {
-        isControlCommandMessage: overrides.isControlCommandMessage ?? (() => false),
-        shouldComputeCommandAuthorized: overrides.shouldComputeCommandAuthorized ?? (() => false),
-        shouldHandleTextCommands: overrides.shouldHandleTextCommands ?? (() => false),
-      },
-      debounce: {
-        resolveInboundDebounceMs: () => overrides.inboundDebounceMs ?? 0,
-        createInboundDebouncer:
-          overrides.createInboundDebouncer ??
-          (<T>(params: { onFlush: (entries: T[]) => Promise<void> | void }) => ({
-            enqueue: async (entry: T) => {
-              await params.onFlush([entry]);
-            },
-          })),
-      },
-      groups: {
-        resolveRequireMention: (params: { requireMentionOverride?: boolean }) =>
-          params.requireMentionOverride ?? false,
-      },
-      media: {
-        readRemoteMediaBuffer: vi.fn(),
-        saveMediaBuffer: vi.fn(),
-      },
-      mentions: {
-        buildMentionRegexes: () => [],
-        matchesMentionPatterns: () => false,
-      },
-      pairing: {
-        buildPairingReply: () => "pairing required",
-      },
-      reply: {
-        settleReplyDispatcher: vi.fn(async ({ onSettled }) => onSettled?.()),
-      },
-      routing: {
-        resolveAgentRoute: () => ({
-          accountId: routeOverride?.accountId ?? "default",
-          agentId: routeOverride?.agentId ?? "main",
-          lastRoutePolicy: routeOverride?.lastRoutePolicy ?? "main",
-          mainSessionKey: routeOverride?.mainSessionKey ?? "mattermost:default:channel:chan-1",
-          sessionKey: routeOverride?.sessionKey ?? "mattermost:default:channel:chan-1",
-        }),
-      },
-      session: {
-        resolveStorePath: () => "/tmp/openclaw-test-sessions.json",
-        recordInboundSession,
-        updateLastRoute: vi.fn(async () => {}),
-      },
-      inbound: {
-        run,
-      },
-      text: {
-        chunkMarkdownTextWithMode:
-          overrides.chunkMarkdownTextWithMode ?? ((text: string) => [text]),
-        convertMarkdownTables: (text: string) => text,
-        hasControlCommand: overrides.textHasControlCommand ?? (() => false),
-        resolveChunkMode: () => overrides.chunkMode ?? "length",
-        resolveMarkdownTableMode: () => "off",
-        resolveTextChunkLimit: () => overrides.textChunkLimit ?? 4000,
-      },
-    },
-  };
-}
-
-const testConfig: OpenClawConfig = {
-  channels: {
-    mattermost: {
-      enabled: true,
-      baseUrl: "https://mattermost.example.com",
-      botToken: "test-auth-token",
-      chatmode: "onmessage",
-      dmPolicy: "open",
-      groupPolicy: "open",
-    },
-  },
-};
-
-const mentionConfig: OpenClawConfig = {
-  ...testConfig,
-  channels: {
-    mattermost: {
-      ...testConfig.channels?.mattermost,
-      chatmode: "oncall",
-    },
-  },
-};
-
-vi.mock("../runtime.js", () => ({
-  getMattermostRuntime: () => mockState.runtimeCore,
-}));
-
-const testRuntime = (): RuntimeEnv =>
-  ({
-    log: vi.fn(),
-    error: vi.fn(),
-    exit: ((code: number): never => {
-      throw new Error(`exit ${code}`);
-    }) as RuntimeEnv["exit"],
-  }) satisfies RuntimeEnv;
-
-async function emitMattermostChannelPost(
-  socket: FakeWebSocket,
-  params: { id: string; message: string; rootId?: string; userId?: string },
-) {
-  const userId = params.userId ?? "user-1";
-  await socket.emitMessage({
-    event: "posted",
-    data: {
-      channel_id: "chan-1",
-      channel_name: "town-square",
-      channel_display_name: "Town Square",
-      sender_name: userId === "user-1" ? "alice" : "bob",
-      post: JSON.stringify({
-        id: params.id,
-        channel_id: "chan-1",
-        user_id: userId,
-        message: params.message,
-        root_id: params.rootId,
-        create_at: 1_714_000_000_000,
-      }),
-    },
-    broadcast: {
-      channel_id: "chan-1",
-      user_id: userId,
-    },
-  });
-}
-
-async function emitMattermostEditedPost(
-  socket: FakeWebSocket,
-  params: {
-    id: string;
-    message: string;
-    editAt: number;
-    channelId?: string;
-    userId?: string;
-    fileIds?: string[];
-    props?: Record<string, unknown>;
-  },
-) {
-  const channelId = params.channelId ?? "chan-1";
-  const userId = params.userId ?? "user-1";
-  await socket.emitMessage({
-    event: "post_edited",
-    data: {
-      channel_id: channelId,
-      channel_name: "town-square",
-      channel_display_name: "Town Square",
-      sender_name: userId === "user-1" ? "alice" : "mallory",
-      post: JSON.stringify({
-        id: params.id,
-        channel_id: channelId,
-        user_id: userId,
-        message: params.message,
-        file_ids: params.fileIds,
-        props: params.props,
-        create_at: 1_714_000_000_000,
-        edit_at: params.editAt,
-      }),
-    },
-    broadcast: {
-      channel_id: channelId,
-      user_id: userId,
-    },
-  });
-}
+const mockState = getMattermostInboundTestState();
 
 describe("mattermost inbound user posts", () => {
-  beforeEach(() => {
-    vi.clearAllMocks();
-    mockState.abortController = undefined;
-    mockState.runtimeCore = createRuntimeCore(testConfig);
-    mockState.createMattermostClient.mockReturnValue({});
-    mockState.createMattermostDraftStream.mockReturnValue({
-      update: vi.fn(),
-      updateAssistantText: vi.fn(),
-      flush: vi.fn(async () => {}),
-      stop: vi.fn(async () => {}),
-      settleBoundaries: vi.fn(async () => {}),
-      resolveFinalText: (text: string) => ({ kind: "full" as const, text }),
-    });
-    mockState.fetchMattermostMe.mockResolvedValue({
-      id: "bot-user",
-      username: "openclaw",
-      update_at: 1,
-    });
-    mockState.registerMattermostMonitorSlashCommands.mockResolvedValue(undefined);
-    mockState.registerPluginHttpRoute.mockReturnValue(vi.fn());
-    mockState.resolveChannelInfo.mockResolvedValue({
-      id: "chan-1",
-      name: "town-square",
-      display_name: "Town Square",
-      team_id: "team-1",
-      type: "O",
-    });
-    mockState.resolveMattermostMedia.mockResolvedValue([]);
-    mockState.resolveUserInfo.mockResolvedValue({ id: "user-1", username: "alice" });
-    mockState.sendMessageMattermost.mockResolvedValue({});
-    mockState.dispatchInboundMessage.mockImplementation(async () => {
-      mockState.abortController?.abort();
-    });
-  });
+  beforeEach(resetMattermostInboundTestState);
 
   it("does not enqueue regular user posts as system events", async () => {
     const socket = new FakeWebSocket();
@@ -1130,657 +631,6 @@ describe("mattermost inbound user posts", () => {
     expect(ctx?.CommandAuthorized).toBe(true);
   });
 
-  it("dispatches an edit that adds a bot mention with an edit-specific message id", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    mockState.runtimeCore = createRuntimeCore(mentionConfig);
-
-    const monitor = monitorMattermostProvider({
-      config: mentionConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-lone-edit",
-      message: "not addressed to the bot",
-    });
-    expect(mockState.dispatchReplyFromConfig).not.toHaveBeenCalled();
-
-    await socket.emitMessage({
-      event: "post_edited",
-      data: {
-        channel_id: "chan-1",
-        channel_name: "town-square",
-        channel_display_name: "Town Square",
-        sender_name: "alice",
-        post: JSON.stringify({
-          id: "post-lone-edit",
-          channel_id: "chan-1",
-          user_id: "user-1",
-          message: "@openclaw edited wake",
-          create_at: 1_714_000_000_000,
-          edit_at: 1_714_000_000_125,
-        }),
-      },
-      broadcast: {
-        channel_id: "chan-1",
-        user_id: "user-1",
-      },
-    });
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-    const ctx = mockState.dispatchReplyFromConfig.mock.calls.at(0)?.[0].ctx;
-    expect(ctx?.BodyForAgent).toBe("edited wake");
-    expect(ctx?.MessageSid).toBe("post-lone-edit");
-    expect(ctx?.MessageSids).toBeUndefined();
-  });
-
-  it("does not dispatch an edited post without a bot mention", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    mockState.runtimeCore = createRuntimeCore(mentionConfig);
-
-    const monitor = monitorMattermostProvider({
-      config: mentionConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await socket.emitMessage({
-      event: "post_edited",
-      data: {
-        channel_id: "chan-1",
-        channel_name: "town-square",
-        channel_display_name: "Town Square",
-        sender_name: "alice",
-        post: JSON.stringify({
-          id: "post-edit-no-mention",
-          channel_id: "chan-1",
-          user_id: "user-1",
-          message: "still not addressed to the bot",
-          create_at: 1_714_000_000_000,
-          edit_at: 1_714_000_000_125,
-        }),
-      },
-      broadcast: {
-        channel_id: "chan-1",
-        user_id: "user-1",
-      },
-    });
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).not.toHaveBeenCalled();
-  });
-
-  it("does not wake again when an edited post already mentioned the bot", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    mockState.runtimeCore = createRuntimeCore(mentionConfig);
-
-    const monitor = monitorMattermostProvider({
-      config: mentionConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-existing-mention",
-      message: "@openclaw original wake",
-    });
-    expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-
-    await emitMattermostEditedPost(socket, {
-      id: "post-existing-mention",
-      message: "@openclaw reformatted original wake",
-      editAt: 1_714_000_000_125,
-    });
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-  });
-
-  it("does not wake for an attachment or metadata-only edit", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    mockState.runtimeCore = createRuntimeCore(mentionConfig);
-
-    const monitor = monitorMattermostProvider({
-      config: mentionConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-metadata-edit",
-      message: "@openclaw original wake",
-    });
-    expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-
-    await emitMattermostEditedPost(socket, {
-      id: "post-metadata-edit",
-      message: "@openclaw original wake",
-      editAt: 1_714_000_000_125,
-      fileIds: ["file-1"],
-      props: { attachments: [{ text: "updated" }] },
-    });
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-  });
-
-  it("does not wake for an edit that adds a different bot mention", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    mockState.runtimeCore = createRuntimeCore(mentionConfig);
-
-    const monitor = monitorMattermostProvider({
-      config: mentionConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-other-bot",
-      message: "not addressed to a bot",
-    });
-    await emitMattermostEditedPost(socket, {
-      id: "post-other-bot",
-      message: "@otherbot edited wake",
-      editAt: 1_714_000_000_125,
-    });
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).not.toHaveBeenCalled();
-  });
-
-  it("fails closed when an edited mention changes the observed sender identity", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    mockState.runtimeCore = createRuntimeCore(mentionConfig);
-
-    const monitor = monitorMattermostProvider({
-      config: mentionConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-sender-change",
-      message: "not addressed to the bot",
-    });
-    await emitMattermostEditedPost(socket, {
-      id: "post-sender-change",
-      message: "@openclaw forged edit",
-      editAt: 1_714_000_000_125,
-      userId: "user-2",
-    });
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).not.toHaveBeenCalled();
-  });
-
-  it("ignores malformed and partial edited-post events", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    mockState.runtimeCore = createRuntimeCore(mentionConfig);
-
-    const monitor = monitorMattermostProvider({
-      config: mentionConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await socket.emitMessage({ event: "post_edited", data: {} });
-    await socket.emitMessage({
-      event: "post_edited",
-      data: {
-        channel_id: "chan-1",
-        post: JSON.stringify({
-          channel_id: "chan-1",
-          user_id: "user-1",
-          message: "@openclaw missing post id",
-        }),
-      },
-    });
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).not.toHaveBeenCalled();
-  });
-
-  it("deduplicates repeated edited-post events", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    mockState.runtimeCore = createRuntimeCore(mentionConfig);
-    mockState.dispatchReplyFromConfig.mockResolvedValue(undefined);
-
-    const monitor = monitorMattermostProvider({
-      config: mentionConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-edit-duplicate",
-      message: "not addressed to the bot",
-    });
-    const editedEvent = {
-      event: "post_edited",
-      data: {
-        channel_id: "chan-1",
-        channel_name: "town-square",
-        channel_display_name: "Town Square",
-        sender_name: "alice",
-        post: JSON.stringify({
-          id: "post-edit-duplicate",
-          channel_id: "chan-1",
-          user_id: "user-1",
-          message: "@openclaw edited wake",
-          create_at: 1_714_000_000_000,
-          edit_at: 1_714_000_000_125,
-        }),
-      },
-      broadcast: {
-        channel_id: "chan-1",
-        user_id: "user-1",
-      },
-    };
-    await socket.emitMessage(editedEvent);
-    await socket.emitMessage(editedEvent);
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-  });
-
-  it("flushes a pending ordinary post before evaluating its edit", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    const debounceConfig: OpenClawConfig = {
-      ...testConfig,
-      messages: { inbound: { debounceMs: 5 } },
-    };
-    const runtimeCore = createRuntimeCore(debounceConfig, undefined, {
-      inboundDebounceMs: 5,
-      createInboundDebouncer,
-    });
-    mockState.runtimeCore = runtimeCore;
-
-    const monitor = monitorMattermostProvider({
-      config: debounceConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await socket.emitMessage({
-      event: "posted",
-      data: {
-        channel_id: "chan-1",
-        channel_name: "town-square",
-        channel_display_name: "Town Square",
-        sender_name: "alice",
-        post: JSON.stringify({
-          id: "post-edited",
-          channel_id: "chan-1",
-          user_id: "user-1",
-          message: "stale pre-edit text",
-          create_at: 1_714_000_000_000,
-        }),
-      },
-      broadcast: {
-        channel_id: "chan-1",
-        user_id: "user-1",
-      },
-    });
-    await socket.emitMessage({
-      event: "post_edited",
-      data: {
-        channel_id: "chan-1",
-        channel_name: "town-square",
-        channel_display_name: "Town Square",
-        sender_name: "alice",
-        post: JSON.stringify({
-          id: "post-edited",
-          channel_id: "chan-1",
-          user_id: "user-1",
-          message: "latest edited text",
-          create_at: 1_714_000_000_000,
-          edit_at: 1_714_000_000_050,
-        }),
-      },
-      broadcast: {
-        channel_id: "chan-1",
-        user_id: "user-1",
-      },
-    });
-
-    await vi.waitFor(() => {
-      expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-    });
-    socket.emitClose(1000);
-    await monitor;
-
-    const ctx = mockState.dispatchReplyFromConfig.mock.calls.at(0)?.[0].ctx;
-    expect(ctx?.BodyForAgent).toBe("stale pre-edit text");
-    expect(ctx?.Body).toContain("stale pre-edit text");
-    expect(ctx?.Body).not.toContain("latest edited text");
-    expect(ctx?.MessageSid).toBe("post-edited");
-    expect(ctx?.MessageSids).toBeUndefined();
-    expect(ctx?.MessageSidFirst).toBeUndefined();
-    expect(ctx?.MessageSidLast).toBeUndefined();
-  });
-
-  it("does not merge a rejected edit into an ordinary debounce batch", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    const debounceConfig: OpenClawConfig = {
-      ...testConfig,
-      messages: { inbound: { debounceMs: 5 } },
-    };
-    mockState.runtimeCore = createRuntimeCore(debounceConfig, undefined, {
-      inboundDebounceMs: 5,
-      createInboundDebouncer,
-    });
-
-    const monitor = monitorMattermostProvider({
-      config: debounceConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-a",
-      message: "@openclaw original post A",
-    });
-    await vi.waitFor(() => {
-      expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-    });
-
-    await emitMattermostChannelPost(socket, { id: "post-b", message: "post B" });
-    await emitMattermostEditedPost(socket, {
-      id: "post-a",
-      message: "@openclaw reformatted post A",
-      editAt: 1_714_000_000_075,
-    });
-
-    await vi.waitFor(() => {
-      expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(2);
-    });
-    socket.emitClose(1000);
-    await monitor;
-
-    const ctx = mockState.dispatchReplyFromConfig.mock.calls.at(1)?.[0].ctx;
-    expect(ctx?.BodyForAgent).toBe("post B");
-    expect(ctx?.Body).not.toContain("reformatted post A");
-    expect(ctx?.MessageSid).toBe("post-b");
-    expect(ctx?.MessageSids).toBeUndefined();
-  });
-
-  it("retains the observed mention state for each debounced sender", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    const debounceConfig: OpenClawConfig = {
-      ...mentionConfig,
-      messages: { inbound: { debounceMs: 5 } },
-    };
-    mockState.runtimeCore = createRuntimeCore(debounceConfig, undefined, {
-      inboundDebounceMs: 5,
-      createInboundDebouncer,
-    });
-
-    const monitor = monitorMattermostProvider({
-      config: debounceConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-alice",
-      message: "alice draft",
-      userId: "user-1",
-    });
-    await emitMattermostChannelPost(socket, {
-      id: "post-bob",
-      message: "bob draft",
-      userId: "user-2",
-    });
-    await emitMattermostEditedPost(socket, {
-      id: "post-alice",
-      message: "@openclaw alice edited wake",
-      editAt: 1_714_000_000_075,
-      userId: "user-1",
-    });
-
-    await vi.waitFor(() => {
-      expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-    });
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    const ctx = mockState.dispatchReplyFromConfig.mock.calls.at(0)?.[0].ctx;
-    expect(ctx?.BodyForAgent).toBe("alice edited wake");
-    expect(ctx?.MessageSid).toBe("post-alice");
-  });
-
-  it("does not wake when an edit adds only an on-character prefix", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    const oncharConfig: OpenClawConfig = {
-      ...testConfig,
-      channels: {
-        mattermost: {
-          ...testConfig.channels?.mattermost,
-          chatmode: "onchar",
-          oncharPrefixes: ["!"],
-        },
-      },
-    };
-    mockState.runtimeCore = createRuntimeCore(oncharConfig);
-
-    const monitor = monitorMattermostProvider({
-      config: oncharConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await emitMattermostChannelPost(socket, {
-      id: "post-onchar-edit",
-      message: "ordinary draft",
-    });
-    await emitMattermostEditedPost(socket, {
-      id: "post-onchar-edit",
-      message: "! edited wake",
-      editAt: 1_714_000_000_075,
-    });
-    abortController.abort();
-    socket.emitClose(1000);
-    await monitor;
-
-    expect(mockState.dispatchReplyFromConfig).not.toHaveBeenCalled();
-  });
-
-  it("does not let bot edited draft posts consume debounced user posts", async () => {
-    const socket = new FakeWebSocket();
-    const abortController = new AbortController();
-    mockState.abortController = abortController;
-    const debounceConfig: OpenClawConfig = {
-      ...testConfig,
-      messages: { inbound: { debounceMs: 5 } },
-    };
-    const runtimeCore = createRuntimeCore(debounceConfig, undefined, {
-      inboundDebounceMs: 5,
-      createInboundDebouncer,
-    });
-    mockState.runtimeCore = runtimeCore;
-
-    const monitor = monitorMattermostProvider({
-      config: debounceConfig,
-      runtime: testRuntime(),
-      abortSignal: abortController.signal,
-      webSocketFactory: () => socket,
-    });
-
-    await vi.waitFor(() => {
-      expect(socket.openListenerCount).toBeGreaterThan(0);
-    });
-    socket.emitOpen();
-
-    await socket.emitMessage({
-      event: "posted",
-      data: {
-        channel_id: "chan-1",
-        channel_name: "town-square",
-        channel_display_name: "Town Square",
-        sender_name: "alice",
-        post: JSON.stringify({
-          id: "post-user",
-          channel_id: "chan-1",
-          user_id: "user-1",
-          message: "user question",
-          create_at: 1_714_000_000_000,
-        }),
-      },
-      broadcast: {
-        channel_id: "chan-1",
-        user_id: "user-1",
-      },
-    });
-    await socket.emitMessage({
-      event: "post_edited",
-      data: {
-        channel_id: "chan-1",
-        channel_name: "town-square",
-        channel_display_name: "Town Square",
-        sender_name: "openclaw",
-        post: JSON.stringify({
-          id: "preview-post",
-          channel_id: "chan-1",
-          user_id: "bot-user",
-          message: "@openclaw bot draft preview",
-          create_at: 1_714_000_000_025,
-          edit_at: 1_714_000_000_050,
-        }),
-      },
-      broadcast: {
-        channel_id: "chan-1",
-        user_id: "bot-user",
-      },
-    });
-
-    await vi.waitFor(() => {
-      expect(mockState.dispatchReplyFromConfig).toHaveBeenCalledTimes(1);
-    });
-    socket.emitClose(1000);
-    await monitor;
-
-    const ctx = mockState.dispatchReplyFromConfig.mock.calls.at(0)?.[0].ctx;
-    expect(ctx?.BodyForAgent).toBe("user question");
-    expect(ctx?.Body).toContain("user question");
-    expect(ctx?.Body).not.toContain("bot draft preview");
-    expect(ctx?.MessageSid).toBe("post-user");
-    expect(ctx?.MessageSids).toBeUndefined();
-  });
-
   it("pins direct-message main route updates to the configured owner", async () => {
     const socket = new FakeWebSocket();
     const abortController = new AbortController();
@@ -2366,4 +1216,3 @@ describe("mattermost inbound user posts", () => {
     );
   });
 });
-/* oxlint-disable max-lines -- TODO: split this grandfathered oversized file. */
